disect pipeline to different functions and add possibility to skip visual analysis at the end of the pipeline. Changed variable names.
